Some groundwork for denoising (#27)
* Type safe handles for resource pool

* Implement stable textures

* Temporal accumulation

* Tweak to string_builder API

* Implemented separable blur node

Refactored string_builder to act more like a builder

* Add quick way of enabling debug printf in shaders

* Bug fixing

* Bug fix for printf include detection

* Small fixes

* Fix build on clang

+++ oblo/vulkan/renderer/src/oblo/vulkan/graph/resource_pool.cpp
@@ -1,10 +1,31 @@
 #include <oblo/vulkan/graph/resource_pool.hpp>
 
 #include <oblo/core/array_size.hpp>
+#include <oblo/core/reflection/struct_compare.hpp>
+#include <oblo/core/reflection/struct_hash.hpp>
 #include <oblo/vulkan/buffer.hpp>
 #include <oblo/vulkan/error.hpp>
 #include <oblo/vulkan/monotonic_gbu_buffer.hpp>
 #include <oblo/vulkan/vulkan_context.hpp>
+
+// For the sake of pooling textures, we ignore the debug labels
+template <>
+struct oblo::equal_to<oblo::debug_label>
+{
+    bool operator()(const oblo::debug_label&, const oblo::debug_label&) const
+    {
+        return true;
+    }
+};
+
+template <>
+struct std::hash<oblo::debug_label>
+{
+    size_t operator()(const oblo::debug_label&) const
+    {
+        return 0;
+    }
+};
 
 namespace oblo::vk
 {
@@ -25,7 +46,7 @@ namespace oblo::vk
             case VK_FORMAT_D16_UNORM_S8_UINT:
             case VK_FORMAT_D24_UNORM_S8_UINT:
             case VK_FORMAT_D32_SFLOAT_S8_UINT:
-                // These have the tencil bit as well, but we cannot create a view for both
+                // These have the stencil bit as well, but we cannot create a view for both
                 // see VUID-VkDescriptorImageInfo-imageView-01976
                 return VK_IMAGE_ASPECT_DEPTH_BIT;
 
@@ -33,6 +54,32 @@ namespace oblo::vk
                 return VK_IMAGE_ASPECT_COLOR_BIT;
             }
         }
+
+        VkImageView create_image_view_2d(
+            VkDevice device, VkImage image, VkFormat format, const VkAllocationCallbacks* allocationCbs)
+        {
+            VkImageView imageView;
+
+            const VkImageViewCreateInfo imageViewInit{
+                .sType = VK_STRUCTURE_TYPE_IMAGE_VIEW_CREATE_INFO,
+                .image = image,
+                .viewType = VK_IMAGE_VIEW_TYPE_2D,
+                .format = format,
+                .subresourceRange =
+                    {
+                        .aspectMask = deduce_aspect_mask(format),
+                        .baseMipLevel = 0,
+                        .levelCount = 1,
+                        .baseArrayLayer = 0,
+                        .layerCount = 1,
+                    },
+            };
+
+            OBLO_VK_PANIC(vkCreateImageView(device, &imageViewInit, allocationCbs, &imageView));
+            return imageView;
+        }
+
+        constexpr u32 FramesBeforeDeletingStableTextures{1};
     }
 
     struct resource_pool::texture_resource
@@ -42,6 +89,8 @@ namespace oblo::vk
         VkImage image;
         VkImageView imageView;
         VkDeviceSize size;
+        h32<stable_texture_resource> stableId;
+        u32 framesAlive;
     };
 
     struct resource_pool::buffer_resource
@@ -111,8 +160,9 @@ namespace oblo::vk
     void resource_pool::shutdown(vulkan_context& ctx)
     {
         m_lastFrameAllocation = m_allocation;
-        std::swap(m_lastFrameTextureResources, m_textureResources);
+        std::swap(m_lastFrameTransientTextures, m_textureResources);
         free_last_frame_resources(ctx);
+        free_stable_textures(ctx, 0);
 
         for (auto& pool : m_bufferPools)
         {
@@ -122,9 +172,9 @@ namespace oblo::vk
 
     void resource_pool::begin_build()
     {
-        m_graphBegin = 0;
+        ++m_frame;
 
-        std::swap(m_lastFrameTextureResources, m_textureResources);
+        std::swap(m_lastFrameTransientTextures, m_textureResources);
         m_textureResources.clear();
 
         m_lastFrameAllocation = m_allocation;
@@ -146,23 +196,25 @@ namespace oblo::vk
 
         create_textures(ctx);
         create_buffers(ctx);
+
+        free_stable_textures(ctx, FramesBeforeDeletingStableTextures);
     }
 
-    void resource_pool::begin_graph()
-    {
-        m_graphBegin = u32(m_textureResources.size());
-    }
-
-    void resource_pool::end_graph() {}
-
-    u32 resource_pool::add(const image_initializer& initializer, lifetime_range range)
+    h32<transient_texture_resource> resource_pool::add_transient_texture(
+        const image_initializer& initializer, lifetime_range range, h32<stable_texture_resource> stableId)
     {
         const auto id = u32(m_textureResources.size());
-        m_textureResources.emplace_back(initializer, range);
-        return id;
+
+        m_textureResources.push_back({
+            .initializer = initializer,
+            .range = range,
+            .stableId = stableId,
+        });
+
+        return h32<transient_texture_resource>{id + 1};
     }
 
-    u32 resource_pool::add_buffer(u32 size, VkBufferUsageFlags usage)
+    h32<transient_buffer_resource> resource_pool::add_transient_buffer(u32 size, VkBufferUsageFlags usage)
     {
         const auto id = u32(m_bufferResources.size());
 
@@ -171,22 +223,24 @@ namespace oblo::vk
             .usage = usage,
         });
 
-        return id;
+        return h32<transient_buffer_resource>{id + 1};
     }
 
-    void resource_pool::add_usage(u32 poolIndex, VkImageUsageFlags usage)
+    void resource_pool::add_transient_texture_usage(h32<transient_texture_resource> transientTexture,
+        VkImageUsageFlags usage)
     {
-        m_textureResources[poolIndex].initializer.usage |= usage;
+        m_textureResources[transientTexture.value - 1].initializer.usage |= usage;
     }
 
-    void resource_pool::add_buffer_usage(u32 poolIndex, VkBufferUsageFlags usage)
+    void resource_pool::add_transient_buffer_usage(h32<transient_buffer_resource> transientBuffer,
+        VkBufferUsageFlags usage)
     {
-        m_bufferResources[poolIndex].usage |= usage;
+        m_bufferResources[transientBuffer.value - 1].usage |= usage;
     }
 
-    texture resource_pool::get_texture(u32 id) const
+    texture resource_pool::get_transient_texture(h32<transient_texture_resource> id) const
     {
-        auto& resource = m_textureResources[id];
+        auto& resource = m_textureResources[id.value - 1];
 
         return {
             .image = resource.image,
@@ -195,9 +249,9 @@ namespace oblo::vk
         };
     }
 
-    buffer resource_pool::get_buffer(u32 id) const
+    buffer resource_pool::get_transient_buffer(h32<transient_buffer_resource> id) const
     {
-        auto& resource = m_bufferResources[id];
+        auto& resource = m_bufferResources[id.value - 1];
 
         return {
             .buffer = resource.buffer,
@@ -206,12 +260,29 @@ namespace oblo::vk
         };
     }
 
+    u32 resource_pool::get_frames_alive_count(h32<transient_texture_resource> id) const
+    {
+        auto& resource = m_textureResources[id.value - 1];
+        return resource.framesAlive;
+    }
+
+    const image_initializer& resource_pool::get_initializer(h32<transient_texture_resource> id) const
+    {
+        auto& resource = m_textureResources[id.value - 1];
+        return resource.initializer;
+    }
+
     void resource_pool::free_last_frame_resources(vulkan_context& ctx)
     {
         const auto submitIndex = ctx.get_submit_index() - 1;
 
-        for (const auto& resource : m_lastFrameTextureResources)
+        for (const auto& resource : m_lastFrameTransientTextures)
         {
+            if (resource.stableId)
+            {
+                continue;
+            }
+
             ctx.destroy_deferred(resource.image, submitIndex);
             ctx.destroy_deferred(resource.imageView, submitIndex);
         }
@@ -234,6 +305,12 @@ namespace oblo::vk
         // For now we just allocate all textures
         for (auto& textureResource : m_textureResources)
         {
+            if (textureResource.stableId)
+            {
+                acquire_from_pool(ctx, textureResource);
+                continue;
+            }
+
             const auto& initializer = textureResource.initializer;
             OBLO_ASSERT(initializer.memoryUsage == memory_usage::gpu_only);
 
@@ -281,27 +358,16 @@ namespace oblo::vk
         VkDeviceSize offset{0};
         for (auto& textureResource : m_textureResources)
         {
+            if (textureResource.stableId)
+            {
+                continue;
+            }
+
             OBLO_VK_PANIC(allocator.bind_image_memory(textureResource.image, m_allocation, offset));
             offset += textureResource.size + textureResource.size % newRequirements.alignment;
 
-            const VkFormat format{textureResource.initializer.format};
-
-            const VkImageViewCreateInfo imageViewInit{
-                .sType = VK_STRUCTURE_TYPE_IMAGE_VIEW_CREATE_INFO,
-                .image = textureResource.image,
-                .viewType = VK_IMAGE_VIEW_TYPE_2D,
-                .format = format,
-                .subresourceRange =
-                    {
-                        .aspectMask = deduce_aspect_mask(format),
-                        .baseMipLevel = 0,
-                        .levelCount = 1,
-                        .baseArrayLayer = 0,
-                        .layerCount = 1,
-                    },
-            };
-
-            OBLO_VK_PANIC(vkCreateImageView(device, &imageViewInit, allocationCbs, &textureResource.imageView));
+            textureResource.imageView =
+                create_image_view_2d(device, textureResource.image, textureResource.initializer.format, allocationCbs);
         }
     }
 
@@ -329,4 +395,72 @@ namespace oblo::vk
             buffer.size = r.size;
         }
     }
+
+    struct resource_pool::stable_texture
+    {
+        allocated_image allocatedImage;
+        VkImageView imageView;
+        u32 creationTime;
+        u32 lastUsedTime;
+    };
+
+    bool resource_pool::stable_texture_key::operator==(const stable_texture_key& rhs) const
+    {
+        return stableId == rhs.stableId && struct_compare<equal_to>(initializer, rhs.initializer);
+    }
+
+    usize resource_pool::stable_texture_key_hash::operator()(const stable_texture_key& key) const
+    {
+        return struct_hash<std::hash>(key);
+    }
+
+    void resource_pool::acquire_from_pool(vulkan_context& ctx, texture_resource& resource)
+    {
+        OBLO_ASSERT(resource.stableId);
+        const auto [it, isNew] =
+            m_stableTextures.try_emplace(stable_texture_key{resource.stableId, resource.initializer});
+
+        if (isNew)
+        {
+            auto& allocator = ctx.get_allocator();
+            OBLO_VK_PANIC(allocator.create_image(resource.initializer, &it->second.allocatedImage));
+
+            it->second.imageView = create_image_view_2d(ctx.get_device(),
+                it->second.allocatedImage.image,
+                resource.initializer.format,
+                allocator.get_allocation_callbacks());
+
+            it->second.creationTime = m_frame;
+        }
+
+        resource.image = it->second.allocatedImage.image;
+        resource.imageView = it->second.imageView;
+        resource.framesAlive = m_frame - it->second.creationTime;
+
+        it->second.lastUsedTime = m_frame;
+    }
+
+    void resource_pool::free_stable_textures(vulkan_context& ctx, u32 unusedFor)
+    {
+        for (auto it = m_stableTextures.begin(); it != m_stableTextures.end();)
+        {
+            const auto dt = m_frame - it->second.lastUsedTime;
+
+            if (dt < unusedFor)
+            {
+                ++it;
+            }
+            else
+            {
+                const auto submitIndex = ctx.get_submit_index();
+
+                ctx.destroy_deferred(it->second.imageView, submitIndex);
+                ctx.destroy_deferred(it->second.allocatedImage.image, submitIndex);
+                ctx.destroy_deferred(it->second.allocatedImage.allocation, submitIndex);
+
+                it = m_stableTextures.erase(it);
+            }
+        }
+    }
+
 }
